+++ projects/org.graalvm.vm.x86/src/org/graalvm/vm/x86/node/debug/trace/Record.java
@@ -95,7 +95,7 @@ public abstract class Record {
         lastStateSupplier = null;
     }
 
-    public long getTid() {
+    public int getTid() {
         return tid;
     }
